+++ message_parser/src/test/java/cumulus_message_adapter/message_parser/AdapterUtilities.java
@@ -21,8 +21,6 @@ import java.util.Map;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
-import com.google.gson.Gson;
-
 /**
  * Utilities for downloading and cleaning up Cumulus Message Adapter package
  */
@@ -89,8 +87,7 @@ public class AdapterUtilities {
                 : CMA_GITHUB_PATH_URL;
 
         String jsonString = getJsonResponse(url);
-        Gson gson = new Gson();
-        Map map = gson.fromJson(jsonString, Map.class);
+        Map<String, Object> map = JsonUtils.toMap(jsonString);
         return (String) map.get("tag_name");
     }
 
@@ -209,35 +206,12 @@ public class AdapterUtilities {
     }
 
     /**
-     * Convert json string to Map object
-     *
-     * @param jsonString The json string
-     * @return The converted Map object
-     */
-    public static Map convertJsonStringToMap(String jsonString) {
-        Gson gson = new Gson();
-        Map expectedOutputJson = gson.fromJson(jsonString, Map.class);
-        return expectedOutputJson;
-    }
-
-    /**
-     * Convert Map object to json string
-     *
-     * @param map The Map object
-     * @return The converted json string
-     */
-    public static String convertMapToJsonString(Map map) {
-        Gson gson = new Gson();
-        return gson.toJson(map);
-    }
-
-    /**
      * load the example output json message from file and update it with TestTask output
      */
-    public static Map getExpectedTestTaskOutputJson() throws IOException
+    public static Map<String, Object> getExpectedTestTaskOutputJson() throws IOException
     {
         String expectedJsonString = loadResourceToString("basic.output.json");
-        Map expectedOutputJson = convertJsonStringToMap(expectedJsonString);
+        Map<String, Object> expectedOutputJson = JsonUtils.toMap(expectedJsonString);
         HashMap<String, String> taskMap = new HashMap<String, String>();
         taskMap.put("task", "complete");
         expectedOutputJson.put("payload", taskMap);