+++ message_parser/src/test/java/cumulus_message_adapter/message_parser/WorkflowException.java
@@ -1,8 +1,13 @@
 package cumulus_message_adapter.message_parser;
 
 /** WorkflowException for testing purposes */
-public class WorkflowException extends Exception 
+public class WorkflowException extends Exception
 {
+    /**
+     *
+     */
+    private static final long serialVersionUID = -1304608916194514402L;
+
     public WorkflowException(String message)
     {
         super(message);