+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [v1.2.13] - 2020-01-02
+
+### Added
+
+- Added `JsonUtils` class to help with converting from a JSON string to a `Map` with proper type assignment
+  - `JsonUtils.toMap()` converts JSON string to a `Map`
+
 ## [v1.2.12] - 2019-12-17
 
 ### Changed